+++ database/migrations/2023_10_04_061548_assignments.php
@@ -13,32 +13,34 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('assignments', function (Blueprint $table) {
-            $table->id();
-            $table->string('sch_id');
-            $table->string('campus');
-            $table->string('period');
-            $table->string('term');
-            $table->string('session');
-            $table->string('teacher_id');
-            $table->string('question_type');
-            $table->longText('question');
-            $table->longText('question_number');
-            $table->longText('answer');
-            $table->unsignedBigInteger('subject_id');
-            $table->string('option1')->nullable();
-            $table->string('option2')->nullable();
-            $table->string('option3')->nullable();
-            $table->string('option4')->nullable();
-            $table->string('image')->nullable();
-            $table->string('total_question')->nullable();
-            $table->string('question_mark')->nullable();
-            $table->string('total_mark')->nullable();
-            $table->string('week')->nullable();
+        if (!Schema::hasTable('assignments')) {
+            Schema::create('assignments', function (Blueprint $table) {
+                $table->id();
+                $table->string('sch_id');
+                $table->string('campus');
+                $table->string('period');
+                $table->string('term');
+                $table->string('session');
+                $table->string('teacher_id');
+                $table->string('question_type');
+                $table->longText('question');
+                $table->longText('question_number');
+                $table->longText('answer');
+                $table->unsignedBigInteger('subject_id');
+                $table->string('option1')->nullable();
+                $table->string('option2')->nullable();
+                $table->string('option3')->nullable();
+                $table->string('option4')->nullable();
+                $table->string('image')->nullable();
+                $table->string('total_question')->nullable();
+                $table->string('question_mark')->nullable();
+                $table->string('total_mark')->nullable();
+                $table->string('week')->nullable();
 
-            $table->foreign('subject_id')->references('id')->on('subjects')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('subject_id')->references('id')->on('subjects')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
