+++ database/migrations/2024_04_23_062822_create_cbt_questions_table.php
@@ -11,34 +11,36 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('cbt_questions', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('cbt_setting_id');
-            $table->string('teacher_id')->nullable();
-            $table->string('sch_id');
-            $table->string('campus');
-            $table->string('period');
-            $table->string('term');
-            $table->string('session');
-            $table->string('subject_id');
-            $table->string('question_type');
-            $table->string('question');
-            $table->string('option1');
-            $table->string('option2');
-            $table->string('option3');
-            $table->string('option4');
-            $table->string('answer');
-            $table->string('question_mark');
-            $table->string('total_mark')->nullable();
-            $table->string('question_number')->nullable();
-            $table->string('total_question')->nullable();
-            $table->string('status')->nullable();
+        if (!Schema::hasTable('cbt_questions')) {
+            Schema::create('cbt_questions', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('cbt_setting_id');
+                $table->string('teacher_id')->nullable();
+                $table->string('sch_id');
+                $table->string('campus');
+                $table->string('period');
+                $table->string('term');
+                $table->string('session');
+                $table->string('subject_id');
+                $table->string('question_type');
+                $table->string('question');
+                $table->string('option1');
+                $table->string('option2');
+                $table->string('option3');
+                $table->string('option4');
+                $table->string('answer');
+                $table->string('question_mark');
+                $table->string('total_mark')->nullable();
+                $table->string('question_number')->nullable();
+                $table->string('total_question')->nullable();
+                $table->string('status')->nullable();
 
-            $table->index(['cbt_setting_id', 'subject_id']);
+                $table->index(['cbt_setting_id', 'subject_id']);
 
-            $table->foreign('cbt_setting_id')->references('id')->on('cbt_settings')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('cbt_setting_id')->references('id')->on('cbt_settings')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
