+++ database/migrations/2024_10_13_064437_create_flip_class_assessment_marks_table.php
@@ -11,30 +11,32 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('flip_class_assessment_marks', function (Blueprint $table) {
-            $table->id();
-            $table->string('sch_id');
-            $table->string('campus');
-            $table->string('period');
-            $table->string('term');
-            $table->string('session');
-            $table->unsignedBigInteger('flip_class_assessment_id');
-            $table->string('student_id');
-            $table->string('subject_id');
-            $table->string('topic');
-            $table->longText('question');
-            $table->string('question_type');
-            $table->string('question_number');
-            $table->longText('answer');
-            $table->longText('correct_answer');
-            $table->string('mark');
-            $table->string('teacher_mark');
-            $table->string('submitted');
-            $table->string('week');
+        if (!Schema::hasTable('flip_class_assessment_marks')) {
+            Schema::create('flip_class_assessment_marks', function (Blueprint $table) {
+                $table->id();
+                $table->string('sch_id');
+                $table->string('campus');
+                $table->string('period');
+                $table->string('term');
+                $table->string('session');
+                $table->unsignedBigInteger('flip_class_assessment_id');
+                $table->string('student_id');
+                $table->string('subject_id');
+                $table->string('topic');
+                $table->longText('question');
+                $table->string('question_type');
+                $table->string('question_number');
+                $table->longText('answer');
+                $table->longText('correct_answer');
+                $table->string('mark');
+                $table->string('teacher_mark');
+                $table->string('submitted');
+                $table->string('week');
 
-            $table->foreign('flip_class_assessment_id')->references('id')->on('flip_class_assessments')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('flip_class_assessment_id')->references('id')->on('flip_class_assessments')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
